* moved ConnectorFactoryTest.cloudNameFromInstanceNameTest to jar-connector; * added es.host:localhost and es.port:12002 properties to pom.xml for tests; * removed dependency on db-testing in db-binding; * moved deprecated ES test DB fixtures from db-serializers to db-testing;
