+++ jar-connector/pom.xml
@@ -147,11 +147,9 @@
 				<artifactId>maven-surefire-plugin</artifactId>
 				<configuration>
 					<systemPropertyVariables>
-						<property>
-							<name>slipstream.config.dir</name>
-							<value>${project.build.directory}/configuration
-							</value>
-						</property>
+						<slipstream.config.dir>${project.build.directory}/configuration</slipstream.config.dir>
+						<es.host>localhost</es.host>
+						<es.port>12002</es.port>
 					</systemPropertyVariables>
 					<useSystemClassLoader>false</useSystemClassLoader>
 					<excludes>
