+++ jar-service/pom.xml
@@ -324,6 +324,8 @@
 				<configuration>
 					<systemPropertyVariables>
 						<slipstream.config.dir>${project.build.directory}/configuration</slipstream.config.dir>
+						<es.host>localhost</es.host>
+            <es.port>12002</es.port>
 					</systemPropertyVariables>
 					<useSystemClassLoader>false</useSystemClassLoader>
 					<includes>
